+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## UNRELEASED
 * Add html class name to resource overview
+* Clear cached session openstad user when making api call or logging out
 
 ## v0.10.1 (2020-12-09)
 * Cache openstad user longer when logged in to cms